Convert GeoJSON-like objects to WKT

Convert GeoJSON-like objects to WKT

geojson2wkt(obj, fmt = 16, third = "z", ...)

Arguments

obj
(list/json/character) A GeoJSON-like object representing a Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon, or GeometryCollection

fmt
Format string which indicates the number of digits to display after the decimal point when formatting coordinates. Max: 20

third
(character) Only applicable when there are three dimensions. If m, assign a M value for a measurement, and if z assign a Z value for three-dimenionsal system. Case is ignored. An M value represents a measurement, while a Z value usually represents altitude (but can be something like depth in a water based location).

...
Further args passed on to jsonlite::fromJSON() only in the event of json passed as a character string (can also be json of class json as returned from jsonlite::toJSON() or simply coerced to json by adding the class manually)

Inputs

Input to obj parameter can take two forms:

A list with named elements type and coordinates OR type and geometries (only in the case of GeometryCollection). e.g., list(type = "Point", coordinates = c(1, 0))

A list with single named element in the set Point, Multipoint, Polygon, Multipolygon, Linestring, Multilinestring,or Geometrycollection, e.g., list(Point = c(1, 0)) - Note that this format is not proper GeoJSON, but is less verbose than the previous format, so should save the user time and make it easier to use.

Each point

For any one point, 2 to 4 values can be used:

2 values: longitude, latitude

3 values: longitude, latitude, altitude

4 values: longitude, latitude, altitude, measure

The 3rd value is typically altitude though can be depth in an aquatic context.

The 4th value is a measurement of some kind.

The GeoJSON spec https://tools.ietf.org/html/rfc7946 actually doesn't allow a 4th value for a point, but we allow it here since we're converting to WKT which does allow a 4th value for a point.

Coordinates data formats

Coordinates data should follow the following formats:

Point: a vector or list, with a single point (2-4 values)

MultiPoint: a matrix, with N points

Linestring: a matrix, with N points

MultiLinestring: the top most level is a list, containing N matrices

Polygon: the top most level is a list, containing N matrices

MultiPolygon: the top most level is a list, the next level is N lists, each of them containing N matrices

Geometrycollection: a list containing any combination and number of the above types

Matrices by definition can not have unequal lengths in their columns, so we don't have to check for that user error.

Each matrix can have any number of rows, and from 2 to 4 columns. If > 5 columns we stop with an error message.

References

https://tools.ietf.org/html/rfc7946, https://en.wikipedia.org/wiki/Well-known_text

See also

Examples

# point
## new format
point <- list(Point = c(116.4, 45.2))
geojson2wkt(point)

#> [1] "POINT (116.4000000000000057  45.2000000000000028)"

## old format, warns
point <- list(type = 'Point', coordinates = c(116.4, 45.2))
geojson2wkt(point)

#> [1] "POINT (116.4000000000000057  45.2000000000000028)"

# multipoint
## new format
mp <- list(MultiPoint = matrix(c(100, 101, 3.14, 3.101, 2.1, 2.18),
   ncol = 2))
geojson2wkt(mp)

#> [1] "MULTIPOINT ((100.0000000000000000 3.1010000000000000), (101.0000000000000000 2.1000000000000001), (3.1400000000000001 2.1800000000000002))"

## 3D
mp <- list(MultiPoint = matrix(c(100, 101, 3, 3, 2, 2, 4, 5, 6),
   ncol = 3))
geojson2wkt(mp)

#> [1] "MULTIPOINT Z((100.0000000000000000 3.0000000000000000 4.0000000000000000), (101.0000000000000000 2.0000000000000000 5.0000000000000000), (3.0000000000000000 2.0000000000000000 6.0000000000000000))"

## old format, warns
mp <- list(
  type = 'MultiPoint',
  coordinates = matrix(c(100, 101, 3.14, 3.101, 2.1, 2.18), ncol = 2)
)
geojson2wkt(mp)

#> [1] "MULTIPOINT ((100.0000000000000000 3.1010000000000000), (101.0000000000000000 2.1000000000000001), (3.1400000000000001 2.1800000000000002))"

# linestring
## new format
st <- list(LineString = matrix(c(0.0, 2.0, 4.0, 5.0,
                               0.0, 1.0, 2.0, 4.0),
                               ncol = 2))
geojson2wkt(st)

#> [1] "LINESTRING (0.0000000000000000 0.0000000000000000, 2.0000000000000000 1.0000000000000000, 4.0000000000000000 2.0000000000000000, 5.0000000000000000 4.0000000000000000)"

## 3D
st <- list(LineString = matrix(
  c(0.0, 0, 0, 
   2, 1, 5,
   100, 300, 800), nrow = 3))
geojson2wkt(st, fmt = 2)

#> [1] "LINESTRING Z(0.00 2.00 100.00, 0.00 1.00 300.00, 0.00 5.00 800.00)"

geojson2wkt(st, fmt = 2, third = "m")

#> [1] "LINESTRING M(0.00 2.00 100.00, 0.00 1.00 300.00, 0.00 5.00 800.00)"

## old format, warns
st <- list(
  type = 'LineString',
  coordinates = matrix(c(0.0, 2.0, 4.0, 5.0,
                         0.0, 1.0, 2.0, 4.0), ncol = 2)
)
geojson2wkt(st)

#> [1] "LINESTRING (0.0000000000000000 0.0000000000000000, 2.0000000000000000 1.0000000000000000, 4.0000000000000000 2.0000000000000000, 5.0000000000000000 4.0000000000000000)"

## 3D
st <- list(LineString = matrix(c(0.0, 2.0, 4.0, 5.0,
                               0.0, 1.0, 2.0, 4.0,
                               10, 20, 30, 40),
                               ncol = 3))
geojson2wkt(st, fmt = 2)

#> [1] "LINESTRING Z(0.00 0.00 10.00, 2.00 1.00 20.00, 4.00 2.00 30.00, 5.00 4.00 40.00)"

## 4D
st <- list(LineString = matrix(c(0.0, 2.0, 4.0, 5.0,
                               0.0, 1.0, 2.0, 4.0,
                               10, 20, 30, 40,
                               1, 2, 3, 4),
                               ncol = 4))
geojson2wkt(st, fmt = 2)

#> [1] "LINESTRING ZM(0.00 0.00 10.00 1.00, 2.00 1.00 20.00 2.00, 4.00 2.00 30.00 3.00, 5.00 4.00 40.00 4.00)"

# multilinestring
## new format
multist <- list(MultiLineString = list(
   matrix(c(0, -2, -4, -1, -3, -5), ncol = 2),
   matrix(c(1.66, 10.9999, 10.9, 0, -31.5, 3.0, 1.1, 0), ncol = 2)
 )
)
geojson2wkt(multist)

#> [1] "MULTILINESTRING ((0.0000000000000000 -1.0000000000000000, -2.0000000000000000 -3.0000000000000000, -4.0000000000000000 -5.0000000000000000), (1.6599999999999999 -31.5000000000000000, 10.9999000000000002 3.0000000000000000, 10.9000000000000004 1.1000000000000001, 0.0000000000000000 0.0000000000000000))"

## 3D
multist <- list(MultiLineString = list(
   matrix(c(0, -2, -4, -1, -3, -5, 100, 200, 300), ncol = 3),
   matrix(c(1, 10, 10.9, 0, -31.5, 3.0, 1.1, 0, 3, 3, 3, 3), ncol = 3)
 )
)
geojson2wkt(multist, fmt = 2)

#> [1] "MULTILINESTRING Z((0.00 -1.00 100.00, -2.00 -3.00 200.00, -4.00 -5.00 300.00), (1.00 -31.50 3.00, 10.00 3.00 3.00, 10.90 1.10 3.00, 0.00 0.00 3.00))"

geojson2wkt(multist, fmt = 2, third = "m")

#> [1] "MULTILINESTRING M((0.00 -1.00 100.00, -2.00 -3.00 200.00, -4.00 -5.00 300.00), (1.00 -31.50 3.00, 10.00 3.00 3.00, 10.90 1.10 3.00, 0.00 0.00 3.00))"

## old format, warns
multist <- list(
  type = 'MultiLineString',
  coordinates = list(
   matrix(c(0, -2, -4, -1, -3, -5), ncol = 2),
   matrix(c(1.66, 10.9999, 10.9, 0, -31.5, 3.0, 1.1, 0), ncol = 2)
 )
)
geojson2wkt(multist)

#> [1] "MULTILINESTRING ((0.0000000000000000 -1.0000000000000000, -2.0000000000000000 -3.0000000000000000, -4.0000000000000000 -5.0000000000000000), (1.6599999999999999 -31.5000000000000000, 10.9999000000000002 3.0000000000000000, 10.9000000000000004 1.1000000000000001, 0.0000000000000000 0.0000000000000000))"

## points within MultiLineString that differ 
## -> use length of longest 
## -> fill with zeros
# 3D and 2D
multist <- list(MultiLineString = list(
    matrix(1:6, ncol = 3), matrix(1:8, ncol = 2)))
geojson2wkt(multist, fmt = 0)

#> [1] "MULTILINESTRING Z((1 3 5, 2 4 6), (1 5 0, 2 6 0, 3 7 0, 4 8 0))"

# 4D and 2D
multist <- list(MultiLineString = list(
    matrix(1:8, ncol = 4), matrix(1:8, ncol = 2)))
geojson2wkt(multist, fmt = 0)

#> [1] "MULTILINESTRING ZM((1 3 5 7, 2 4 6 8), (1 5 0 0, 2 6 0 0, 3 7 0 0, 4 8 0 0))"

# 2D and 2D
multist <- list(MultiLineString = list(
    matrix(1:4, ncol = 2), matrix(1:8, ncol = 2)))
geojson2wkt(multist, fmt = 0)

#> [1] "MULTILINESTRING ((1 3, 2 4), (1 5, 2 6, 3 7, 4 8))"

# 5D and 2D - FAILS
# multist <- list(MultiLineString = list(
#    matrix(1:10, ncol = 5), matrix(1:8, ncol = 2)))
# geojson2wkt(multist, fmt = 0)

# polygon
## new format
poly <- list(Polygon = list(
   matrix(c(100.001, 101.1, 101.001, 100.001, 0.001, 0.001, 1.001, 0.001), ncol = 2),
   matrix(c(100.201, 100.801, 100.801, 100.201, 0.201, 0.201, 0.801, 0.201), ncol = 2)
))
geojson2wkt(poly)

#> [1] "POLYGON ((100.0010000000000048 0.0010000000000000, 101.0999999999999943 0.0010000000000000, 101.0010000000000048 1.0009999999999999, 100.0010000000000048 0.0010000000000000), (100.2009999999999934 0.2010000000000000, 100.8010000000000019 0.2010000000000000, 100.8010000000000019 0.8010000000000000, 100.2009999999999934 0.2010000000000000))"

geojson2wkt(poly, fmt=6)

#> [1] "POLYGON ((100.001000 0.001000, 101.100000 0.001000, 101.001000 1.001000, 100.001000 0.001000), (100.201000 0.201000, 100.801000 0.201000, 100.801000 0.801000, 100.201000 0.201000))"

## 3D
poly <- list(Polygon = list(
   matrix(c(100.1, 101.1, 101.1, 100.1, 0.1, 0.1, 1.1, 0.1, 1, 1, 1, 1), ncol = 3),
   matrix(c(100.2, 100.8, 100.8, 100.2, 0.2, 0.2, 0.80, 0.2, 3, 3, 3, 3), ncol = 3)
))
geojson2wkt(poly, fmt = 2)

#> [1] "POLYGON Z((100.10 0.10 1.00, 101.10 0.10 1.00, 101.10 1.10 1.00, 100.10 0.10 1.00), (100.20 0.20 3.00, 100.80 0.20 3.00, 100.80 0.80 3.00, 100.20 0.20 3.00))"

geojson2wkt(poly, fmt = 2, third = "m")

#> [1] "POLYGON M((100.10 0.10 1.00, 101.10 0.10 1.00, 101.10 1.10 1.00, 100.10 0.10 1.00), (100.20 0.20 3.00, 100.80 0.20 3.00, 100.80 0.80 3.00, 100.20 0.20 3.00))"

## old format, warns
poly <- list(
  type = 'Polygon',
  coordinates = list(
    matrix(c(100.001, 101.1, 101.001, 100.001, 0.001, 0.001, 1.001, 0.001),
      ncol = 2),
    matrix(c(100.201, 100.801, 100.801, 100.201, 0.201, 0.201, 0.801, 0.201),
      ncol = 2)
  )
)
geojson2wkt(poly)

#> [1] "POLYGON ((100.0010000000000048 0.0010000000000000, 101.0999999999999943 0.0010000000000000, 101.0010000000000048 1.0009999999999999, 100.0010000000000048 0.0010000000000000), (100.2009999999999934 0.2010000000000000, 100.8010000000000019 0.2010000000000000, 100.8010000000000019 0.8010000000000000, 100.2009999999999934 0.2010000000000000))"

geojson2wkt(poly, fmt=6)

#> [1] "POLYGON ((100.001000 0.001000, 101.100000 0.001000, 101.001000 1.001000, 100.001000 0.001000), (100.201000 0.201000, 100.801000 0.201000, 100.801000 0.801000, 100.201000 0.201000))"

## points within Polygon that differ 
## -> use length of longest 
## -> fill with zeros
# 3D and 2D
poly <- list(Polygon = list(
   matrix(c(100, 101, 101, 100, 0.1, 0.2, 0.3, 0.1, 5, 6, 7, 8), ncol = 3),
   matrix(c(40, 41, 61, 40, 0.1, 0.2, 0.3, 0.1), ncol = 2)
))
geojson2wkt(poly, fmt = 0)

#> [1] "POLYGON Z((100.0 0.1 5.0, 101.0 0.2 6.0, 101.0 0.3 7.0, 100.0 0.1 8.0), (40.0 0.1 0.0, 41.0 0.2 0.0, 61.0 0.3 0.0, 40.0 0.1 0.0))"

# 4D and 2D
poly <- list(Polygon = list(
   matrix(c(100, 101, 101, 100, 0.1, 0.2, 0.3, 0.1, 5, 6, 7, 8, 1, 1, 1, 1), 
     ncol = 4),
   matrix(c(40, 41, 61, 40, 0.1, 0.2, 0.3, 0.1), ncol = 2)
))
geojson2wkt(poly, fmt = 0)

#> [1] "POLYGON ZM((100.0 0.1 5.0 1.0, 101.0 0.2 6.0 1.0, 101.0 0.3 7.0 1.0, 100.0 0.1 8.0 1.0), (40.0 0.1 0.0 0.0, 41.0 0.2 0.0 0.0, 61.0 0.3 0.0 0.0, 40.0 0.1 0.0 0.0))"

# 5D and 2D - FAILS
# multist <- list(Polygon = list(
#    matrix(1:10, ncol = 5), matrix(1:8, ncol = 2)))
# geojson2wkt(poly, fmt = 0)

# multipolygon
## new format
mpoly <- list(MultiPolygon = list(
  list(
    matrix(c(100, 101, 101, 100, 0.001, 0.001, 1.001, 0.001), ncol = 2),
    matrix(c(100.2, 100.8, 100.8, 100.2, 0.2, 0.2, 0.8, 0.2), ncol = 2)
  ),
  list(
    matrix(c(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 1.0), ncol = 2),
    matrix(c(9.0, 10.0, 11.0, 12.0, 1.0, 2.0, 3.0, 4.0, 9.0), ncol = 2)
  )
))

#> Warning: data length [9] is not a sub-multiple or multiple of the number of rows [5]

#> Warning: data length [9] is not a sub-multiple or multiple of the number of rows [5]

geojson2wkt(mpoly, fmt=2)

#> [1] "MULTIPOLYGON (((100.000 0.001, 101.000 0.001, 101.000 1.001, 100.000 0.001), (100.20 0.20, 100.80 0.20, 100.80 0.80, 100.20 0.20)), ((1.00 6.00, 2.00 7.00, 3.00 8.00, 4.00 1.00, 5.00 1.00), (9.00 2.00, 10.00 3.00, 11.00 4.00, 12.00 9.00, 1.00 9.00)))"

## 3D
mpoly <- list(MultiPolygon = list(
  list(
    matrix(c(100, 101, 101, 100, 0.001, 0.001, 1.001, 0.001, 1, 1, 1, 1), 
      ncol = 3),
    matrix(c(100.2, 100.8, 100.8, 100.2, 0.2, 0.2, 0.8, 0.2, 3, 4, 5, 6), 
      ncol = 3)
  ),
  list(
    matrix(c(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 1.0, 1, 1, 1, 1),
      ncol = 3),
    matrix(c(9.0, 10.0, 11.0, 12.0, 1.0, 2.0, 3.0, 4.0, 9.0, 9, 9, 9, 9),
      ncol = 3)
  )
))

#> Warning: data length [13] is not a sub-multiple or multiple of the number of rows [5]

#> Warning: data length [13] is not a sub-multiple or multiple of the number of rows [5]

geojson2wkt(mpoly, fmt=2)

#> [1] "MULTIPOLYGON Z(((100.000 0.001 1.000, 101.000 0.001 1.000, 101.000 1.001 1.000, 100.000 0.001 1.000), (100.20 0.20 3.00, 100.80 0.20 4.00, 100.80 0.80 5.00, 100.20 0.20 6.00)), ((1.00 6.00 1.00, 2.00 7.00 1.00, 3.00 8.00 1.00, 4.00 1.00 1.00, 5.00 1.00 2.00), (9.00 2.00 9.00, 10.00 3.00 9.00, 11.00 4.00 9.00, 12.00 9.00 9.00, 1.00 9.00 10.00)))"

geojson2wkt(mpoly, fmt=2, third = "m")

#> [1] "MULTIPOLYGON M(((100.000 0.001 1.000, 101.000 0.001 1.000, 101.000 1.001 1.000, 100.000 0.001 1.000), (100.20 0.20 3.00, 100.80 0.20 4.00, 100.80 0.80 5.00, 100.20 0.20 6.00)), ((1.00 6.00 1.00, 2.00 7.00 1.00, 3.00 8.00 1.00, 4.00 1.00 1.00, 5.00 1.00 2.00), (9.00 2.00 9.00, 10.00 3.00 9.00, 11.00 4.00 9.00, 12.00 9.00 9.00, 1.00 9.00 10.00)))"

## old format, warns
mpoly <- list(
  type = 'MultiPolygon',
  coordinates = list(
    list(
      matrix(c(100, 101, 101, 100, 0.001, 0.001, 1.001, 0.001), ncol = 2),
      matrix(c(100.2, 100.8, 100.8, 100.2, 0.2, 0.2, 0.8, 0.2), ncol = 2)
    ),
    list(
      matrix(c(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 1.0), ncol = 3),
      matrix(c(9.0, 10.0, 11.0, 12.0, 1.0, 2.0, 3.0, 4.0, 9.0), ncol = 3)
    )
  )
)
geojson2wkt(mpoly, fmt=2)

#> [1] "MULTIPOLYGON Z(((100.000 0.001 0.000, 101.000 0.001 0.000, 101.000 1.001 0.000, 100.000 0.001 0.000), (100.20 0.20 0.00, 100.80 0.20 0.00, 100.80 0.80 0.00, 100.20 0.20 0.00)), ((1.00 4.00 7.00, 2.00 5.00 8.00, 3.00 6.00 1.00), (9.00 12.00 3.00, 10.00 1.00 4.00, 11.00 2.00 9.00)))"

mpoly2 <- list(
  type = "MultiPolygon",
  coordinates = list(
    list(list(c(30, 20), c(45, 40), c(10, 40), c(30, 20))),
    list(list(c(15, 5), c(40, 10), c(10, 20), c(5 ,10), c(15, 5)))
  )
)

mpoly2 <- list(
  type = "MultiPolygon",
  coordinates = list(
    list(
      matrix(c(30, 45, 10, 30, 20, 40, 40, 20), ncol = 2)
    ),
    list(
      matrix(c(15, 40, 10, 5, 15, 5, 10, 20, 10, 5), ncol = 2)
    )
  )
)
geojson2wkt(mpoly2, fmt=1)

#> [1] "MULTIPOLYGON (((30.0 20.0, 45.0 40.0, 10.0 40.0, 30.0 20.0)), ((15.0 5.0, 40.0 10.0, 10.0 20.0, 5.0 10.0, 15.0 5.0)))"

## points within MultiPolygon that differ 
## -> use length of longest 
## -> fill with zeros
# 3D and 2D
mpoly <- list(MultiPolygon = list(
  list(
    matrix(c(40, 130, 155, 40, 20, 34, 34, 20), ncol = 2),
    matrix(c(30, 40, 54, 30, 0.1, 42, 62, 0.1, 1, 1, 1, 1), ncol = 3)
  ),
  list(
    matrix(c(9, 49, 79, 9, 11, 35, 15, 11), ncol = 2),
    matrix(c(1, 33, 59, 1, 5, 16, 36, 5), ncol = 2)
  )
))
geojson2wkt(mpoly, fmt = 0)

#> [1] "MULTIPOLYGON Z(((40 20 0, 130 34 0, 155 34 0, 40 20 0), (30.0 0.1 1.0, 40.0 42.0 1.0, 54.0 62.0 1.0, 30.0 0.1 1.0)), ((9 11 0, 49 35 0, 79 15 0, 9 11 0), (1 5 0, 33 16 0, 59 36 0, 1 5 0)))"

# 4D and 2D
mpoly <- list(MultiPolygon = list(
  list(
    matrix(c(40, 130, 155, 40, 20, 34, 34, 20), ncol = 2),
    matrix(c(30, 40, 54, 30, 0.1, 42, 62, 0.1, 1, 1, 1, 1, 0, 0, 0, 0), ncol = 4)
  ),
  list(
    matrix(c(9, 49, 79, 9, 11, 35, 15, 11), ncol = 2),
    matrix(c(1, 33, 59, 1, 5, 16, 36, 5), ncol = 2)
  )
))
geojson2wkt(mpoly, fmt = 0)

#> [1] "MULTIPOLYGON ZM(((40 20 0 0, 130 34 0 0, 155 34 0 0, 40 20 0 0), (30.0 0.1 1.0 0.0, 40.0 42.0 1.0 0.0, 54.0 62.0 1.0 0.0, 30.0 0.1 1.0 0.0)), ((9 11 0 0, 49 35 0 0, 79 15 0 0, 9 11 0 0), (1 5 0 0, 33 16 0 0, 59 36 0 0, 1 5 0 0)))"

# 5D and 2D - FAILS
mpoly <- list(MultiPolygon = list(
  list(
    matrix(c(40, 130, 155, 40, 20, 34, 34, 20), ncol = 2),
    matrix(c(30, 40, 54, 30, 
             0.1, 42, 62, 0.1, 
             1, 1, 1, 1, 
             0, 0, 0, 0,
             0, 0, 0, 0), ncol = 5)
  ),
  list(
    matrix(c(9, 49, 79, 9, 11, 35, 15, 11), ncol = 2),
    matrix(c(1, 33, 59, 1, 5, 16, 36, 5), ncol = 2)
  )
))
# geojson2wkt(mpoly, fmt = 0)

# geometrycollection
## new format
gmcoll <- list(GeometryCollection = list(
 list(Point = c(0.0, 1.0)),
 list(LineString = matrix(c(0.0, 2.0, 4.0, 5.0,
                           0.0, 1.0, 2.0, 4.0),
                           ncol = 2)),
 list(Polygon = list(
   matrix(c(100.001, 101.1, 101.001, 100.001, 0.001, 0.001, 1.001, 0.001),
     ncol = 2),
   matrix(c(100.201, 100.801, 100.801, 100.201, 0.201, 0.201, 0.801, 0.201),
     ncol = 2)
  ))
))
geojson2wkt(gmcoll, fmt=0)

#> [1] "GEOMETRYCOLLECTION (POINT (0 1), LINESTRING (0 0, 2 1, 4 2, 5 4), POLYGON ((100.001 0.001, 101.100 0.001, 101.001 1.001, 100.001 0.001), (100.201 0.201, 100.801 0.201, 100.801 0.801, 100.201 0.201)))"

## old format, warns
gmcoll <- list(
 type = 'GeometryCollection',
 geometries = list(
   list(type = 'Point', coordinates = c(0.0, 1.0)),
   list(type = 'LineString', coordinates = matrix(c(0.0, 2.0, 4.0, 5.0,
                           0.0, 1.0, 2.0, 4.0),
                           ncol = 2)),
   list(type = 'Polygon', coordinates = list(
     matrix(c(100.001, 101.1, 101.001, 100.001, 0.001, 0.001, 1.001, 0.001),
       ncol = 2),
     matrix(c(100.201, 100.801, 100.801, 100.201, 0.201, 0.201, 0.801, 0.201),
       ncol = 2)
  ))
 )
)
geojson2wkt(gmcoll, fmt=0)

#> [1] "GEOMETRYCOLLECTION (POINT (0 1), LINESTRING (0 0, 2 1, 4 2, 5 4), POLYGON ((100.001 0.001, 101.100 0.001, 101.001 1.001, 100.001 0.001), (100.201 0.201, 100.801 0.201, 100.801 0.801, 100.201 0.201)))"

# Convert geojson as character string to WKT
# new format
str <- '{
   "Point": [
       -105.01621,
       39.57422
   ]
}'
geojson2wkt(str)

#> [1] "POINT (-105.0162100000000009   39.5742199999999968)"

## old format, warns
str <- '{
   "type": "Point",
   "coordinates": [
       -105.01621,
       39.57422
   ]
}'
geojson2wkt(str)

#> [1] "POINT (-105.0162100000000009   39.5742199999999968)"

## new format
str <- '{"LineString":[[0,0,10],[2,1,20],[4,2,30],[5,4,40]]}'
geojson2wkt(str)

#> [1] "LINESTRING Z(0 0 10, 2 1 20, 4 2 30, 5 4 40)"

## old format, warns
str <-
'{"type":"LineString","coordinates":[[0,0,10],[2,1,20],[4,2,30],[5,4,40]]}'
geojson2wkt(str)

#> [1] "LINESTRING Z(0 0 10, 2 1 20, 4 2 30, 5 4 40)"

# From a jsonlite json object
library("jsonlite")
json <- toJSON(list(Point=c(-105,39)), auto_unbox=TRUE)
geojson2wkt(json)

#> [1] "POINT (-105   39)"

## old format, warns
json <- toJSON(list(type="Point", coordinates=c(-105,39)), auto_unbox=TRUE)
geojson2wkt(json)

#> [1] "POINT (-105   39)"