Make WKT multipoint objects

Make WKT multipoint objects

multipoint(..., fmt = 16, third = "z")

Arguments

...
A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.

fmt
Format string which indicates the number of digits to display after the decimal point when formatting coordinates. Max: 20

third
(character) Only applicable when there are three dimensions. If m, assign a M value for a measurement, and if z assign a Z value for three-dimenionsal system. Case is ignored. An M value represents a measurement, while a Z value usually represents altitude (but can be something like depth in a water based location).

See also

Examples

## empty multipoint
multipoint("empty")

#> [1] "MULTIPOINT EMPTY"

# multipoint("stuff")

# numeric
multipoint(c(100.000, 3.101), c(101.000, 2.100), c(3.140, 2.180))

#> [1] "MULTIPOINT ((100.0000000000000000 3.1010000000000000), (101.0000000000000000 2.1000000000000001), (3.1400000000000001 2.1800000000000002))"

# data.frame
df <- us_cities[1:25, c('long', 'lat')]
multipoint(df)

#> [1] "MULTIPOINT ((-99.7399999999999949 32.4500000000000028), (-81.5199999999999960 41.0799999999999983), (-122.2600000000000051 37.7700000000000031), (-84.1800000000000068 31.5799999999999983), (-73.7999999999999972 42.6700000000000017), (-123.0900000000000034 44.6199999999999974), (-106.6200000000000045 35.1199999999999974), (-92.4599999999999937 31.2899999999999991), (-77.0900000000000034 38.8200000000000003), (-118.1299999999999955 34.0799999999999983), (-117.7300000000000040 33.5700000000000003), (-96.6700000000000017 33.1099999999999994), (-75.4800000000000040 40.6000000000000014), (-122.8700000000000045 45.4900000000000020), (-118.1299999999999955 34.1899999999999977), (-81.4000000000000057 28.6600000000000001), (-78.4000000000000057 40.5099999999999980), (-101.8199999999999932 35.2000000000000028), (-93.6299999999999955 42.0200000000000031), (-117.8700000000000045 33.8400000000000034), (-149.1899999999999977 61.1799999999999997), (-85.6899999999999977 40.0900000000000034), (-71.1400000000000006 42.6499999999999986), (-83.7300000000000040 42.2800000000000011), (-77.2099999999999937 38.8299999999999983))"

# matrix
mat <- matrix(c(df$long, df$lat), ncol = 2)
multipoint(mat)

#> [1] "MULTIPOINT ((-99.7399999999999949 32.4500000000000028), (-81.5199999999999960 41.0799999999999983), (-122.2600000000000051 37.7700000000000031), (-84.1800000000000068 31.5799999999999983), (-73.7999999999999972 42.6700000000000017), (-123.0900000000000034 44.6199999999999974), (-106.6200000000000045 35.1199999999999974), (-92.4599999999999937 31.2899999999999991), (-77.0900000000000034 38.8200000000000003), (-118.1299999999999955 34.0799999999999983), (-117.7300000000000040 33.5700000000000003), (-96.6700000000000017 33.1099999999999994), (-75.4800000000000040 40.6000000000000014), (-122.8700000000000045 45.4900000000000020), (-118.1299999999999955 34.1899999999999977), (-81.4000000000000057 28.6600000000000001), (-78.4000000000000057 40.5099999999999980), (-101.8199999999999932 35.2000000000000028), (-93.6299999999999955 42.0200000000000031), (-117.8700000000000045 33.8400000000000034), (-149.1899999999999977 61.1799999999999997), (-85.6899999999999977 40.0900000000000034), (-71.1400000000000006 42.6499999999999986), (-83.7300000000000040 42.2800000000000011), (-77.2099999999999937 38.8299999999999983))"

# list
multipoint(list(c(100.000, 3.101), c(101.000, 2.100), c(3.140, 2.180)))

#> [1] "MULTIPOINT ((100.0000000000000000 3.1010000000000000), (101.0000000000000000 2.1000000000000001), (3.1400000000000001 2.1800000000000002))"

## a 3rd point is included
multipoint(c(100, 3, 0), c(101, 2, 0), c(3, 2, 0), 
  third = "z", fmt = 1)

#> [1] "MULTIPOINT Z((100.0 3.0 0.0), (101.0 2.0 0.0), (3.0 2.0 0.0))"

multipoint(c(100, 3, 0), c(101, 2, 0), c(3, 2, 0), 
  third = "m", fmt = 1)

#> [1] "MULTIPOINT M((100.0 3.0 0.0), (101.0 2.0 0.0), (3.0 2.0 0.0))"

df <- us_cities[1:25, c('long', 'lat')]
df$altitude <- round(runif(25, 100, 500))
multipoint(df, fmt = 2)

#> [1] "MULTIPOINT Z((-99.74 32.45 299.00), (-81.52 41.08 216.00), (-122.26 37.77 393.00), (-84.18 31.58 409.00), (-73.80 42.67 450.00), (-123.09 44.62 170.00), (-106.62 35.12 114.00), (-92.46 31.29 228.00), (-77.09 38.82 261.00), (-118.13 34.08 178.00), (-117.73 33.57 261.00), (-96.67 33.11 125.00), (-75.48 40.60 255.00), (-122.87 45.49 490.00), (-118.13 34.19 216.00), (-81.40 28.66 371.00), (-78.40 40.51 394.00), (-101.82 35.20 178.00), (-93.63 42.02 492.00), (-117.87 33.84 397.00), (-149.19 61.18 121.00), (-85.69 40.09 312.00), (-71.14 42.65 378.00), (-83.73 42.28 375.00), (-77.21 38.83 112.00))"

multipoint(df, fmt = 2, third = "m")

#> [1] "MULTIPOINT M((-99.74 32.45 299.00), (-81.52 41.08 216.00), (-122.26 37.77 393.00), (-84.18 31.58 409.00), (-73.80 42.67 450.00), (-123.09 44.62 170.00), (-106.62 35.12 114.00), (-92.46 31.29 228.00), (-77.09 38.82 261.00), (-118.13 34.08 178.00), (-117.73 33.57 261.00), (-96.67 33.11 125.00), (-75.48 40.60 255.00), (-122.87 45.49 490.00), (-118.13 34.19 216.00), (-81.40 28.66 371.00), (-78.40 40.51 394.00), (-101.82 35.20 178.00), (-93.63 42.02 492.00), (-117.87 33.84 397.00), (-149.19 61.18 121.00), (-85.69 40.09 312.00), (-71.14 42.65 378.00), (-83.73 42.28 375.00), (-77.21 38.83 112.00))"

mat <- matrix(1:9, 3)
multipoint(mat)

#> [1] "MULTIPOINT Z((1 4 7), (2 5 8), (3 6 9))"

multipoint(mat, third = "m")

#> [1] "MULTIPOINT M((1 4 7), (2 5 8), (3 6 9))"

x <- list(c(100, 3, 0), c(101, 2, 1), c(3, 2, 5))
multipoint(x)

#> [1] "MULTIPOINT Z((100.0000000000000000 3.0000000000000000 0.0000000000000000), (101.0000000000000000 2.0000000000000000 1.0000000000000000), (3.0000000000000000 2.0000000000000000 5.0000000000000000))"

## a 4th point is included
multipoint(
  c(100, 3, 0, 500), c(101, 2, 0, 505), c(3, 2, 0, 600), 
  fmt = 1)

#> [1] "MULTIPOINT ZM((100.0 3.0 0.0 500.0), (101.0 2.0 0.0 505.0), (3.0 2.0 0.0 600.0))"

df <- us_cities[1:25, c('long', 'lat')]
df$altitude <- round(runif(25, 100, 500))
df$weight <- round(runif(25, 1, 100))
multipoint(df, fmt = 2)

#> [1] "MULTIPOINT ZM((-99.74 32.45 190.00 21.00), (-81.52 41.08 220.00 72.00), (-122.26 37.77 355.00 7.00), (-84.18 31.58 292.00 36.00), (-73.80 42.67 273.00 83.00), (-123.09 44.62 383.00 28.00), (-106.62 35.12 479.00 57.00), (-92.46 31.29 172.00 34.00), (-77.09 38.82 187.00 60.00), (-118.13 34.08 372.00 20.00), (-117.73 33.57 300.00 95.00), (-96.67 33.11 357.00 55.00), (-75.48 40.60 364.00 55.00), (-122.87 45.49 138.00 29.00), (-118.13 34.19 406.00 45.00), (-81.40 28.66 408.00 38.00), (-78.40 40.51 496.00 4.00), (-101.82 35.20 488.00 47.00), (-93.63 42.02 256.00 40.00), (-117.87 33.84 284.00 3.00), (-149.19 61.18 226.00 38.00), (-85.69 40.09 170.00 56.00), (-71.14 42.65 313.00 86.00), (-83.73 42.28 297.00 39.00), (-77.21 38.83 412.00 53.00))"

mat <- matrix(1:12, 3)
multipoint(mat)

#> [1] "MULTIPOINT ZM((1 4 7 10), (2 5 8 11), (3 6 9 12))"

x <- list(c(100, 3, 0, 300), c(101, 2, 1, 200), c(3, 2, 5, 100))
multipoint(x, fmt = 3)

#> [1] "MULTIPOINT ZM((100.000 3.000 0.000 300.000), (101.000 2.000 1.000 200.000), (3.000 2.000 5.000 100.000))"