Make WKT polygon objects

Make WKT polygon objects

polygon(..., fmt = 16, third = "z")

Arguments

...
A GeoJSON-like object representing a Point, LineString, Polygon, MultiPolygon, etc.

fmt
Format string which indicates the number of digits to display after the decimal point when formatting coordinates. Max: 20

third
(character) Only applicable when there are three dimensions. If m, assign a M value for a measurement, and if z assign a Z value for three-dimenionsal system. Case is ignored. An M value represents a measurement, while a Z value usually represents altitude (but can be something like depth in a water based location).

Details

You can create nested polygons with list and data.frame inputs, but not from numeric inputs. See examples.

See also

Examples

## empty polygon
polygon("empty")

#> [1] "POLYGON EMPTY"

# polygon("stuff")

# numeric
polygon(c(100.001, 0.001), c(101.12345, 0.001), c(101.001, 1.001),
  c(100.001, 0.001), fmt=2)

#> [1] "POLYGON ((100.001 0.001, 101.1235 0.0010, 101.001 1.001, 100.001 0.001))"

# data.frame
## single polygon
df <- us_cities[2:5,c('long','lat')]
df <- rbind(df, df[1,])
wktview(polygon(df, fmt=2), zoom=4)
## nested polygons
df2 <- data.frame(long = c(-85.9, -85.9, -93, -93, -85.9),
                  lat = c(37.5, 35.3, 35.3, 37.5, 37.5))
wktview(polygon(df, df2, fmt=2), zoom=4)

# matrix
mat <- matrix(c(df$long, df$lat), ncol = 2)
polygon(mat)

#> [1] "POLYGON ((-81.5199999999999960 41.0799999999999983, -122.2600000000000051 37.7700000000000031, -84.1800000000000068 31.5799999999999983, -73.7999999999999972 42.6700000000000017, -81.5199999999999960 41.0799999999999983))"

# list
# single list - creates single polygon
ply <- list(c(100.001, 0.001), c(101.12345, 0.001), c(101.001, 1.001),
  c(100.001, 0.001))
wktview(polygon(ply, fmt=2), zoom=7)
# nested list - creates nested polygon
vv <- polygon(list(c(35, 10), c(45, 45), c(15, 40), c(10, 20), c(35, 10)),
   list(c(20, 30), c(35, 35), c(30, 20), c(20, 30)), fmt=2)
wktview(vv, zoom=3)
# multiple lists nested within a list
zz <- polygon(list(list(c(35, 10), c(45, 45), c(15, 40), c(10, 20), c(35, 10)),
   list(c(20, 30), c(35, 35), c(30, 20), c(20, 30))), fmt=2)
wktview(zz, zoom=3)

## a 3rd point is included
### numeric
polygon(c(100, 0, 30), c(101, 0, 30), c(101, 1, 30),
  c(100, 0, 30), fmt = 2)

#> [1] "POLYGON Z((100.00 0.00 30.00, 101.00 0.00 30.00, 101.00 1.00 30.00, 100.00 0.00 30.00))"

polygon(c(100, 0, 30), c(101, 0, 30), c(101, 1, 30),
  c(100, 0, 30), fmt = 2, third = "m")

#> [1] "POLYGON M((100.00 0.00 30.00, 101.00 0.00 30.00, 101.00 1.00 30.00, 100.00 0.00 30.00))"

### data.frame
df <- us_cities[2:5, c('long','lat')]
df <- rbind(df, df[1,])
df$altitude <- round(runif(NROW(df), 10, 50))
polygon(df, fmt=2)

#> [1] "POLYGON Z((-81.52 41.08 34.00, -122.26 37.77 20.00, -84.18 31.58 22.00, -73.80 42.67 29.00, -81.52 41.08 47.00))"

polygon(df, fmt=2, third = "m")

#> [1] "POLYGON M((-81.52 41.08 34.00, -122.26 37.77 20.00, -84.18 31.58 22.00, -73.80 42.67 29.00, -81.52 41.08 47.00))"

### matrix
mat <- matrix(c(df$long, df$lat, df$altitude), ncol = 3)
polygon(mat, fmt=2)

#> [1] "POLYGON Z((-81.52 41.08 34.00, -122.26 37.77 20.00, -84.18 31.58 22.00, -73.80 42.67 29.00, -81.52 41.08 47.00))"

polygon(mat, fmt=2, third = "m")

#> [1] "POLYGON M((-81.52 41.08 34.00, -122.26 37.77 20.00, -84.18 31.58 22.00, -73.80 42.67 29.00, -81.52 41.08 47.00))"

### list
ply <- list(c(100, 0, 1), c(101, 0, 1), c(101, 1, 1),
  c(100, 0, 1))
polygon(ply, fmt=2)

#> [1] "POLYGON Z((100.00 0.00 1.00, 101.00 0.00 1.00, 101.00 1.00 1.00, 100.00 0.00 1.00))"

polygon(ply, fmt=2, third = "m")

#> [1] "POLYGON M((100.00 0.00 1.00, 101.00 0.00 1.00, 101.00 1.00 1.00, 100.00 0.00 1.00))"

## a 4th point is included
### numeric
polygon(c(100, 0, 30, 3.5), c(101, 0, 30, 3.5), c(101, 1, 30, 3.5),
  c(100, 0, 30, 3.5), fmt = 2)

#> [1] "POLYGON ZM((100.00 0.00 30.00 3.50, 101.00 0.00 30.00 3.50, 101.00 1.00 30.00 3.50, 100.00 0.00 30.00 3.50))"

### data.frame
df <- us_cities[2:5, c('long','lat')]
df <- rbind(df, df[1,])
df$altitude <- round(runif(NROW(df), 10, 50))
df$weight <- round(runif(NROW(df), 0, 1), 1)
polygon(df, fmt=2)

#> [1] "POLYGON ZM((-81.52 41.08 26.00 0.10, -122.26 37.77 19.00 0.50, -84.18 31.58 37.00 0.80, -73.80 42.67 12.00 0.70, -81.52 41.08 50.00 0.20))"

### matrix
mat <- matrix(unname(unlist(df)), ncol = 4)
polygon(mat, fmt=2)

#> [1] "POLYGON ZM((-81.52 41.08 26.00 0.10, -122.26 37.77 19.00 0.50, -84.18 31.58 37.00 0.80, -73.80 42.67 12.00 0.70, -81.52 41.08 50.00 0.20))"

### list
ply <- list(c(100, 0, 1, 40), c(101, 0, 1, 44), c(101, 1, 1, 45),
  c(100, 0, 1, 49))
polygon(ply, fmt=2)

#> [1] "POLYGON ZM((100.00 0.00 1.00 40.00, 101.00 0.00 1.00 44.00, 101.00 1.00 1.00 45.00, 100.00 0.00 1.00 49.00))"